refactor(orc8r): Seggregation of Ctraced servicer as external
